requirements.txt
-e .
tabulate
colorama
pandas
nepse-cli
tqdm
